+++ inst/yupana/server.R
@@ -4,7 +4,7 @@
 #> open https://flavjack.github.io/inti/
 #> open https://flavjack.shinyapps.io/yupanapro/
 #> author .: Flavio Lozano-Isla (lozanoisla.com)
-#> date .: 2021-04-29
+#> date .: 2021-05-04
 # -------------------------------------------------------------------------
 
 # -------------------------------------------------------------------------
@@ -169,8 +169,7 @@ observe({
     if ( input$fieldbook_gsheet %in% sheet_names(gs()) ) {
 
       gs() %>%
-        range_read( input$fieldbook_gsheet ) %>% 
-        as.data.frame()
+        range_read( input$fieldbook_gsheet )
 
     } else { fieldbook <- NULL }
 
@@ -192,6 +191,130 @@ observe({
     
   })
 
+# Yupana: Exploratory -----------------------------------------------------
+# -------------------------------------------------------------------------
+  
+  output$raw_response <- renderUI({
+    
+    validate(need(fieldbook(), "LogIn and create or insert a url"))
+    
+    variable_names <- fieldbook() %>%
+      names()
+    
+    selectInput(
+      inputId = "raw_y"
+      , label = "Response"
+      , choices = c("choose" = ""
+                    , variable_names)
+    )
+    
+  })
+  
+  output$raw_x <- renderUI({
+    
+    validate(need(fieldbook(), "LogIn and create or insert a url"))
+    
+    factor_names <- fieldbook() %>%
+      names()
+    
+    selectInput(
+      inputId = "raw_x"
+      , label = "Axis X"
+      , choices = c("choose" = ""
+                    , factor_names)
+    )
+    
+  })
+  
+  output$raw_group <- renderUI({
+    
+    validate(need(fieldbook(), "LogIn and create or insert a url"))
+    
+    factor_names <- fieldbook() %>%
+      names()
+    
+    selectInput(
+      inputId = "raw_group"
+      , label = "Grouped"
+      , choices = c("choose" = ""
+                    , factor_names)
+    )
+    
+  })
+  
+  plotraw <- reactive({
+    
+    validate(need(fieldbook(), "LogIn and create or insert a url"))
+    validate(need(input$raw_x, "Choose your parameters"))
+    validate(need(input$raw_y, "Choose your parameters"))
+    
+    raw_xrotation <- input$raw_xrotation %>% 
+      strsplit(., "[*]") %>% 
+      pluck(1) %>% as.numeric()
+    
+    raw_ylimits <- input$raw_ylimits %>% 
+      strsplit(., "[*]") %>% 
+      pluck(1) %>% as.numeric()
+    
+    raw_gtext <- input$raw_gtext %>% 
+      strsplit(., ",") %>% 
+      pluck(1) %>% as.character()
+    
+    raw_xtext <- input$raw_xtext %>% 
+      strsplit(., ",") %>% 
+      pluck(1) %>% as.character()
+    
+    fieldbook() %>% 
+      plot_raw(type = "boxplot"
+               , x = if(input$raw_x == "") NULL else input$raw_x
+               , y = if(input$raw_y == "") NULL else input$raw_y
+               , group = if(input$raw_group == "") NULL else input$raw_group
+               , xlab = if(input$raw_xlab == "") NULL else input$raw_xlab
+               , ylab = if(input$raw_ylab == "") NULL else input$raw_ylab
+               , glab = if(input$raw_glab == "") NULL else input$raw_glab
+               , ylimits = if(input$raw_ylimits == "") NULL else raw_ylimits
+               , xrotation = if(input$raw_xrotation == "") NULL else raw_xrotation
+               , legend = input$raw_legend
+               , color = if(input$raw_color == "yes") TRUE else FALSE
+               , opt = if(input$raw_opt == "") NULL else input$raw_opt
+               , xtext = if(input$raw_xtext == "") NULL else raw_xtext
+               , gtext = if(input$raw_gtext == "") NULL else raw_gtext
+               )
+    })
+  
+  output$plotraw <- renderImage({
+    
+    validate(need(fieldbook(), "LogIn and create or insert a url"))
+    
+    
+    dim <- input$raw_dimension %>% 
+      strsplit(., "[*]") %>% 
+      pluck(1) %>% as.numeric()
+    
+    if(is.na(dim[1])) { ancho <- 20 } else  { ancho <- dim[1] }
+    if(is.na(dim[2])) { alto <- 10 } else  { alto <- dim[2] }
+    if(is.na(dim[3])) { dpi <- 100 } else  { dpi <- dim[3] }
+    
+    outfile <- tempfile(fileext = ".png")
+    
+    png(outfile, width = ancho, height = alto, units = "cm", res = dpi)
+    print(plotraw())
+    dev.off()
+    
+    list(src = outfile)
+    
+  }, deleteFile = TRUE)
+  
+# ------------------------------------------------------------------------- 
+  
+  output$plot_raw <- renderUI({
+    
+    validate(need(plotraw(), "Choose your parameters"))
+
+    tagList( div(imageOutput("plotraw"), align = "center") ) 
+    
+    })
+
 # Yupana: Fieldbook -------------------------------------------------------
 # -------------------------------------------------------------------------
 
@@ -202,7 +325,7 @@ observe({
 
   })
 
-  # summary module ----------------------------------------------------------
+# summary module ----------------------------------------------------------
 
   output$last_factor <- renderUI({
 
@@ -762,7 +885,7 @@ observe({
   
   plotgr <- reactive({ plot_smr(plotdt()) })
 
-  # -------------------------------------------------------------------------
+# -------------------------------------------------------------------------
 
   output$plotgr <- renderImage({